split main file into separate headers
